+++ EdblockModel/SymbolsModel/LineSymbolsModel/DecoratorLineSymbolsModel/CoordinateDecorator.cs
@@ -2,12 +2,12 @@
 
 public class CoordinateDecorator : ICoordinateDecorator
 {
-    public int X { get; set; }
-    public int Y { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
 
     public CoordinateDecorator() =>
         (X, Y) = (0, 0);
 
-    public CoordinateDecorator((int x, int y) borderCoordinateBlockSymbol) =>
+    public CoordinateDecorator((double x, double y) borderCoordinateBlockSymbol) =>
         (X, Y) = (borderCoordinateBlockSymbol.x, borderCoordinateBlockSymbol.y);
 }
