+++ EdblockViewModel/Symbols/ActionSymbolVM.cs
@@ -55,7 +55,7 @@ public class ActionSymbolVM : BlockSymbolVM, IHaveCommentVM
         SetHeight(Height);
     }
 
-    public override void SetWidth(int width)
+    public override void SetWidth(double width)
     {
         Width = width;
 
@@ -70,7 +70,7 @@ public class ActionSymbolVM : BlockSymbolVM, IHaveCommentVM
         CommentSymbolVM?.SetCoordinateBlockSymbol();
     }
 
-    public override void SetHeight(int height)
+    public override void SetHeight(double height)
     {
         Height = height;
