+++ EdblockViewModel/Symbols/ConditionSymbolVM.cs
@@ -48,7 +48,7 @@ public class ConditionSymbolVM : BlockSymbolVM, IHavePolygon
         SetHeight(Height);
     }
 
-    public override void SetWidth(int width)
+    public override void SetWidth(double width)
     {     
         Width = width;
 
@@ -62,7 +62,7 @@ public class ConditionSymbolVM : BlockSymbolVM, IHavePolygon
         ChangeCoordinateAuxiliaryElements();
     }
 
-    public override void SetHeight(int height)
+    public override void SetHeight(double height)
     {
         Height = height;
