+++ EdblockViewModel/Symbols/CycleWhileEndSymbolVM.cs
@@ -51,7 +51,7 @@ public class CycleWhileEndSymbolVM : BlockSymbolVM, IHavePolygon
         SetHeight(Height);
     }
 
-    public override void SetWidth(int width)
+    public override void SetWidth(double width)
     {
         Width = width;
 
@@ -65,7 +65,7 @@ public class CycleWhileEndSymbolVM : BlockSymbolVM, IHavePolygon
         ChangeCoordinateAuxiliaryElements();
     }
 
-    public override void SetHeight(int height)
+    public override void SetHeight(double height)
     {
         Height = height;
