+++ SerializationEdblock/LineSymbolSerializable.cs
@@ -2,8 +2,8 @@
 
 public class LineSymbolSerializable
 {
-    public int X1 { get; set; }
-    public int Y1 { get; set; }
-    public int X2 { get; set; }
-    public int Y2 { get; set; }
+    public double X1 { get; set; }
+    public double Y1 { get; set; }
+    public double X2 { get; set; }
+    public double Y2 { get; set; }
 }
